+++ lib/nostr.dart
@@ -4,3 +4,5 @@
 library nostr;
 
 export 'src/event.dart';
+export 'src/keychain.dart';
+export 'src/utils.dart';