+++ lib/src/event.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'package:convert/convert.dart';
 import 'package:crypto/crypto.dart';
-import 'package:bip340/bip340.dart';
+import 'package:bip340/bip340.dart' as bip340;
 import 'package:nostr/src/utils.dart';
 
 /// The only object type that exists is the event, which has the following format on the wire:
@@ -29,18 +29,21 @@ class Event {
   /// -  0: set_metadata: the content is set to a stringified JSON object {name: <username>, about: <string>, picture: <url, string>} describing the user who created the event. A relay may delete past set_metadata events once it gets a new one for the same pubkey.
   /// -  1: text_note: the content is set to the text content of a note (anything the user wants to say). Non-plaintext notes should instead use kind 1000-10000 as described in NIP-16.
   /// -  2: recommend_server: the content is set to the URL (e.g., wss://somerelay.com) of a relay the event creator wants to recommend to its followers.
-  int kind;
+  late int kind;
 
   /// The tags array can store a tag identifier as the first element of each subarray, plus arbitrary information afterward (always as strings).
   ///
   /// This NIP defines "p" — meaning "pubkey", which points to a pubkey of someone that is referred to in the event —, and "e" — meaning "event", which points to the id of an event this event is quoting, replying to or referring to somehow.
-  List<List<String>> tags;
+  late List<List<String>> tags;
 
   /// arbitrary string
-  String content;
+  String content = "";
 
   /// 64-bytes signature of the sha256 hash of the serialized event data, which is the same as the "id" field
   late String sig;
+
+  /// subscription_id is a random string that should be used to represent a subscription.
+  String? subscriptionId;
 
   /// Default constructor
   Event(
@@ -53,40 +56,41 @@ class Event {
     this.sig,
   ) {
     assert(createdAt.toString().length == 10);
-    assert(createdAt <= currentUnixSecondsTimestamp());
+    assert(createdAt <= currentUnixTimestampSeconds());
     pubkey = pubkey.toLowerCase();
     String id = getEventId();
     assert(this.id == id);
-    assert(verify(pubkey, id, sig));
+    assert(bip340.verify(pubkey, id, sig));
   }
 
   /// Instanciate Event object from the minimum available data
-  Event.from({
-    this.createdAt = 0,
-    required this.kind,
-    required this.tags,
-    required this.content,
-    required String privkey,
-  }) {
+  Event.from(
+      {this.createdAt = 0,
+      required this.kind,
+      required this.tags,
+      required this.content,
+      required String privkey,
+      this.subscriptionId}) {
     if (createdAt == 0) {
-      createdAt = currentUnixSecondsTimestamp();
+      createdAt = currentUnixTimestampSeconds();
     }
     assert(createdAt.toString().length == 10);
-    assert(createdAt <= currentUnixSecondsTimestamp());
-    pubkey = getPublicKey(privkey).toLowerCase();
+    assert(createdAt <= currentUnixTimestampSeconds());
+    pubkey = bip340.getPublicKey(privkey).toLowerCase();
     id = getEventId();
     sig = getSignature(privkey);
   }
 
   /// Deserialize an event from a JSON
-  Event.fromJson(Map<String, dynamic> json)
-      : id = json['id'],
-        pubkey = json['pubkey'],
-        createdAt = json['created_at'],
-        kind = json['kind'],
-        tags = json['tags'],
-        content = json['content'],
-        sig = json['sig'];
+  Event.fromJson(Map<String, dynamic> json, {this.subscriptionId}) {
+    id = json['id'];
+    pubkey = json['pubkey'];
+    createdAt = json['created_at'];
+    kind = json['kind'];
+    tags = json['tags'].cast<List<String>>();
+    content = json['content'];
+    sig = json['sig'];
+  }
 
   /// Serialize an event in JSON
   Map<String, dynamic> toJson() => {
@@ -100,9 +104,36 @@ class Event {
       };
 
   /// Serialize to nostr event message
-  /// - ["EVENT", event JSON], used to publish events.
+  /// - ["EVENT", event JSON as defined above]
+  /// - ["EVENT", subscription_id, event JSON as defined above]
   String serialize() {
-    return jsonEncode(["EVENT", toJson()]);
+    if (subscriptionId != null) {
+      return jsonEncode(["EVENT", subscriptionId, toJson()]);
+    } else {
+      return jsonEncode(["EVENT", toJson()]);
+    }
+  }
+
+  /// Deserialize a nostr event message
+  /// - ["EVENT", event JSON as defined above]
+  /// - ["EVENT", subscription_id, event JSON as defined above]
+  Event.deserialize(input) {
+    Map<String, dynamic> json = {};
+    if (input.length == 2) {
+      json = input[1] as Map<String, dynamic>;
+    } else if (input.length == 3) {
+      json = input[2] as Map<String, dynamic>;
+      subscriptionId = input[1];
+    } else {
+      throw Exception('invalid input');
+    }
+    id = json['id'];
+    pubkey = json['pubkey'];
+    createdAt = json['created_at'];
+    kind = json['kind'];
+    tags = json['tags'].cast<List<String>>();
+    content = json['content'];
+    sig = json['sig'];
   }
 
   /// To obtain the event.id, we sha256 the serialized event.
@@ -128,7 +159,7 @@ class Event {
   String getSignature(String privateKey) {
     /// aux must be 32-bytes random bytes, generated at signature time.
     /// https://github.com/nbd-wtf/dart-bip340/blob/master/lib/src/bip340.dart#L10
-    String aux = generate32RandomBytes();
-    return sign(privateKey, id, aux);
+    String aux = generate64RandomHexChars();
+    return bip340.sign(privateKey, id, aux);
   }
 }