+++ lib/src/utils.dart
@@ -2,14 +2,14 @@ import 'dart:math';
 
 import 'package:convert/convert.dart';
 
-/// generate 32 random bytes
-String generate32RandomBytes() {
+/// generates 32 random bytes converted in hex
+String generate64RandomHexChars() {
   final random = Random.secure();
   final randomBytes = List<int>.generate(32, (i) => random.nextInt(256));
   return hex.encode(randomBytes);
 }
 
 /// current unix timestamp in seconds
-int currentUnixSecondsTimestamp() {
+int currentUnixTimestampSeconds() {
   return DateTime.now().millisecondsSinceEpoch ~/ 1000;
 }